MDS-638: Key is checked for existence before be written
If check-for-update option is set, proxy will lookup key. If key is
found, proxy will response 403 instead of writing the key. If elliptics
group responses with status, that is not equal to -ENOENT or zero,
proxy will assume, that key can be in that group and also will not allow
to write key.