MDS-670: upload handlers can retry couple
Writing will be retried into another couple if there is unused couple
and key is not found in that unused couple. (Unused means unused during
current request.) The couple, that the writing was failed into, is
pessimized.

Also trick to avoid thevoid's bug was removed due the bag was fixed.