MDS-1617: Prolong writer_t lifetime in method write
We need to prolong the lifetime of the shared state here to be sure it
is alive till the end of the function.
The reason of being uncertain is calling async_result.connect(next_) in
this function below. In that call user's callback may be called
in-place, and the only external shared pointer to the shared state may
be destroyed there.